+++ benches/benchmark_e448_224.rs
@@ -1,8 +1,8 @@
-use capycrypt::ecc::ecc_encryptable::KeyEncryptable;
-use capycrypt::ecc::ecc_keypair::KeyPair;
-use capycrypt::ecc::ecc_signable::Signable;
+use capycrypt::ecc::encryptable::KeyEncryptable;
+use capycrypt::ecc::keypair::KeyPair;
+use capycrypt::ecc::signable::Signable;
 use capycrypt::sha3::aux_functions::byte_utils::get_random_bytes;
-use capycrypt::sha3::sponge_encryptable::SpongeEncryptable;
+use capycrypt::sha3::encryptable::SpongeEncryptable;
 use capycrypt::SecParam::D224;
 use capycrypt::{Message, SecParam};
 use criterion::{criterion_group, criterion_main, Criterion};
