+++ tests/aes_tests.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod aes_modes_tests {
-    use capycrypt::aes::aes_encryptable::AesEncryptable;
+    use capycrypt::aes::encryptable::AesEncryptable;
     use capycrypt::sha3::aux_functions::byte_utils::get_random_bytes;
     use capycrypt::Message;
